requirements.txt
langgraph
tenacity
requests
termcolor
python-dotenv
httpx
nest_asyncio
MainContentExtractor
ratelimit
pyaudio
pyperclip
keyboard
pydantic
mcp
pyGithub
GitPython